% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stregion.R
\name{contorno_circulo}
\alias{contorno_circulo}
\title{Gera um contorno circular (ou arco)}
\usage{
contorno_circulo(
  cx = 0,
  cy = 0,
  r = 1,
  ang_ini = 0,
  ang_fim = 2 * pi,
  n_pts = 200
)
}
\arguments{
\item{cx, cy}{Centro do circulo.}

\item{r}{Raio.}

\item{ang_ini, ang_fim}{Angulos inicial e final em radianos.}

\item{n_pts}{Numero de pontos do poligono aproximador.}
}
\value{
Data frame com colunas \code{x} e \code{y}.
}
\description{
Gera um contorno circular (ou arco)
}
