% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stregion.R
\name{contorno_elipse}
\alias{contorno_elipse}
\title{Gera um contorno eliptico (ou arco de elipse)}
\usage{
contorno_elipse(
  cx = 0,
  cy = 0,
  a = 1,
  b = 0.5,
  ang_ini = 0,
  ang_fim = 2 * pi,
  n_pts = 200
)
}
\arguments{
\item{cx, cy}{Centro.}

\item{a}{Semi-eixo horizontal.}

\item{b}{Semi-eixo vertical.}

\item{ang_ini, ang_fim}{Angulos parametricos inicial e final.}

\item{n_pts}{Numero de pontos.}
}
\value{
Data frame com colunas \code{x} e \code{y}.
}
\description{
Gera um contorno eliptico (ou arco de elipse)
}
