% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stregion.R
\name{contorno_poligono}
\alias{contorno_poligono}
\title{Gera um contorno a partir de vertices arbitrarios}
\usage{
contorno_poligono(vx, vy, n_pts = 200)
}
\arguments{
\item{vx, vy}{Coordenadas dos vertices.}

\item{n_pts}{Numero de pontos ao longo do perimetro.}
}
\value{
Data frame com colunas \code{x} e \code{y}.
}
\description{
Gera um contorno a partir de vertices arbitrarios
}
